+++ src/components/Input/RecurringInput/RecurringInput.scss
@@ -1,0 +1,16 @@
+@use '../../../utilities/colors' as v;
+
+.recurring {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  font-family: 'Inter', sans-serif;
+  background: #ffffff;
+  border: 1px solid v.$accent-gray;
+  border-radius: 1rem;
+  padding: 0.5rem;
+  color: v.$dark-gray;
+  font-weight: 500;
+  font-size: 0.875rem;
+  cursor: pointer;
+}
